Update dependencies
nokogiri has known vulnerabilities, and dependabot seems to be busy
doing PR for updating gem gratuitously rather than doing PR to fix
vulnerabilities when they occur.

Instruct dependabot to stop trying to update gems to reduce the
maintenance burden, and rely on our internal tooling (based on
bundle-audit) to notify us when an update is needed.